掲示板２[レビュー]

+++ board2/delete.php
@@ -6,25 +6,31 @@ $smarty=new MySmarty();
 
 
     try{
-        $db = getDb();
+
         //送られてきたデータを変数に格納
         $id = $_POST['id'];
         $user_id = $_POST['user_id'];
 
-        if($user_id == $_SESSION['login_id']){
-            //削除ボタンを押されたpostデータを削除
-            $stt = $db->prepare("DELETE FROM post WHERE id = :id ");
-            $stt->bindValue(':id', $id);
-            $stt->execute();
+        if(isset($id)&&($user_id)){
 
-            print "削除しました";
-            $db = NULL;
 
+            if ($user_id == $_SESSION['login_id']) {
+                $db = getDb();
+                //削除ボタンを押されたpostデータを削除
+                $stt = $db->prepare("DELETE FROM post WHERE id = :id ");
+                $stt->bindValue(':id', $id);
+                $stt->execute();
+
+                print "削除しました";
+                $db = NULL;
+
+            }
+        }else{
+            print "エラーが発生しました";
         }
-
         
     }catch (PDOException $e) {
         $db = NULL;
         die("エラーメッセージ:{$e->getMessage()}");
-}$smarty->display('edit.php');
+}$smarty->display('smar_delete.tpl');
